Inputs: Avoid implicit conversion of int to float with AxisNotify.
X and Y offsets calculated in OnAxisNotify are of type int.
AxisNotify has the function attributes as type float, thus
passed in attributes get converted to float type. These
values are used in BrowserProcessDispatcherDelegate and
passed to MouseWheelEvent constructor. MouseWheelEvent
function attributes are of type int. This CL fixes these
un-necessary implicit conversions by changing the function
attributes ( e.g. X and Y offset) of OnAxisNotify to type int.

+++ impl/ipc/browser_process_dispatcher_delegate.cc
@@ -59,8 +59,8 @@ void BrowserProcessDispatcherDelegate::ButtonNotify(unsigned handle,
 
 void BrowserProcessDispatcherDelegate::AxisNotify(float x,
                                                   float y,
-                                                  float xoffset,
-                                                  float yoffset) {
+                                                  int xoffset,
+                                                  int yoffset) {
   gfx::Point position(x, y);
   ui::MouseEvent mouseev(ui::ET_MOUSEWHEEL, position, position, 0, 0);
